Fixed missing resistor for VCA CV input
Boards before this commit missed a 200k resistor on the CV input of the VCA,
which caused the VCA to be very hot.

Additionally, we've added a small optional passive LP network to the CV input
to smoothen sudden changes on the CV input (which can cause very minor pops
on very short attack times for example).